+++ .gitignore
@@ -1,4 +1,5 @@
 *.class
+*.iml
 
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
